requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.6
python-dateutil==2.9.0.post0
python-whois==0.9.3
requests==2.31.0
six==1.16.0
urllib3==2.2.1
